eww/.config/eww/dashboard/scripts/mstodo-eww/requirements.txt
msal==1.18.0
python-dotenv==1.0.0
requests==2.28.1
